Started working on the software
For now I'm using the example on Mcp320x library. It is not working but I suspect that this is due to the use of the input buffers: If there is no input, the output from the op amp is unstable. A possible solution is to use a pull down resistor (or pull up) so that it stabilizes op amp output. A large resistor (1M) should work.

This might not be the best solution. Perhaps a better op amp could be used. Since signal conditioning will be used, it might be interesting to eliminate the buffer entirely. It will be much simpler.